smurf: Add typo checking for iterative map-maker config files
We now have default values for the iterative map-maker that are
all defined in one place: smurf_makemap.def.

The config is now read using kpg1Config which handles the
merging of sub-instrument configs, sets the keymap to be
locked (so that only values in the defaults can be
over-ridden) and also sets the keymap to trigger an
error if a keyword is requested that does not exist.

All access to the config keymap has been simplified
since we now guarantee that a value will be available.
We only need to check the return value of astMapGetXX
if the value defaults to <undef> (such as for the csotau
or notch filter).