smurf: Fix up heater values for flatfield ramps
The problem is that the data were assumed to be taken with 3 measurements
in each heater setting. What actually happened was that the first 5 were
done at the reference setting and then the data were grouped in threes
finishing with a single value at the reference setting again.

For example the heater values and the actual values look something like

   Stored    1 1 1 2 2 2 3 3 3 4 4 4 5 5 5
   Actual    0 0 0 0 0 2 2 2 3 3 3 4 4 4 5

So we can correct for this by starting at the end and copying in the value
two slots further down until we get to position #4. Then replacing that with
the PIXHEAT number.