KAPPA: Restore functioning container mode work in PLUCK.
Do not use LPG for NDFs propagated to a container file.  LPG should
access the name of the container file, not the generated NDF name
within the container file.  As previously coded LPG_PROP retained the
original container name, and so the results were sent to the same
dummy NDF that was erased at the end.