smurf: add zerobad option to sc2fft
The reason for this is that model files (like COM) can have VAL__BADD
floating around... which causes problems for FFTs done with the FFTW
library. Simply setting all of these bad values to 0 is a simple way
to get things to work, although a better solution might involve using
the quality arrays to search for the start/end of padding, and perhaps
calling the gap-filling routine (although this latter solution
for small holes within the data can be achieved with sc2clean).