smurf: Allow gain box fit to be performed over a larger box.
The new config parameter "com.fit_box" allows control over the size of
the box over which the least squares linear fit is performed when finding
the gain and offset for each block of time slices in a bolometer (by
comparing the bolometer values with the common mode signal). So you can
set com.fit_box = 6000 and com.gain_box = 100 (say) to get a "rolling"
fit - i.e. a separate gain is estimated for each block of 100 time
slices, but the fit is based on a box of length 6000 time slices (centred
on the 100 slice box).

The default is com.gain_box is 6000. The default for com.fit_box is whatever
was supplied for com.gain_box.

Be warned, having a small com.gain_box involves huge numbers of linear
fits, and so can slow makemap down enormously. In an attempt to
counter-act this, this commit also includes a speed-ip for the estimation
of the common mode signal, in that gains and offsets are no longer
recalculated for blocks of time slices that have already converged.