smurf: Add plane subtraction (PLN) model to iterative map-maker
This is a first cut at a plane subtraction model.

Add "pln" to the model list. I have seen instability when putting
"pln" before "flt".

It is not multi-threaded. It does not fit a plane for all subarrays
simultaneously. It does not calculate a change in chisq for the
model. It has not been tested very much. There are no parameters
to control the fit.

It does use the pointing model for the fit though so it will be
relatively easy to process all subarrays.