smurf: initial commit of new task SC2EXPANDMODEL
The model NDF exported by the map-maker can be difficult to analyze
(in particular the DKS model). The purpose of this task is to expand
an exported model component into a full 3d time-series. To accomplish
this, the FITS headers for models files now have extra information,
such as the 3-character model type (SMFMODEL), and the full dimensions
of the original 3d time-series data cube that the model was fit to
(SMFDIM* and SMFLBND*). smf_open_file now reads the model type into
hdr->mtype so that it is easy to check whether a given file is
a model or not (mtype != SMF__NUL).

Generic routines with names of the form smf_expmodel_??? will then
do the work specific to each model, and function pointers can be
obtained using smf_model_getexpptr(). One difficulty is that we
cannot rely on the model containing WCS for the time axis. So we
now re-generate the time series WCS using sc2store calls combined
with information stored in the JCMTState and generic SCUBA-2
FITS header (both of which are exported to all models) -- see
smf_create_tswcs which wraps sc2store_timeWcs.

Presently things don't quite work since the data are read using
smf_open_file, and it will not read in the JCMTState if the model
does not consist of 3 NDF dimensions. It should be modified so that
JCMTState is read in regardless of the NDF dimensions (and ensure
that the correct number of time slices are recorded in hdr->nframes).