smurf: store dark squids in smfData within smfDA
Dark squids are detector data like any other... it was annoying
having to handle the dks using specialized code all the time.
Now the dark squids are immediated converted to double when they
are loaded, and stored in a 3d smfData (pointed to from smfDA),
but with the row axis having length 1. This should allow us to
use most of the smfData-based routines.