smurf: Do not correct padding when fixing steps
Previously, smf_fix_steps altered the zero values used to pad the start
and end of each bolometer time series, by subtracting off the mean
correction. The common mode signal is later subtracted form the bulk of
the time stream, bringing the values down close to zero, but COM is not
subtracted from the padding. So a large jump in value could be left at
the start of the padding, resulting in bad ringing when the filter model
is applied.