smurf: Improve step correction algorithm
After looking at Remo's data (see his "SC2CLEAN?" message from 21st May),
I have tweaked one or two aspects of the algorithm to find the steps that
were being missed. They were missed because each step consisted of two
separate sub-steps separated by around 50 samples. These sub-steps were
too close together to be processed as separate steps, but when combined
into a single step, they formed a step that was wider than the widest
allowed step, and so were ignored. So I have raised the limit on how wide
a single step can be (dcmaxstepwidth), and also reduced the minimum gap
between steps (dcminstepgap) to encourage the sub-steps to be treated as
separate steps.

I've also change the way the least squares fits are performed that are
used to determine the actual height of each step (potentially in the
presence of a gradient). Previously, each fit was performed over a box of
dcfitbox samples, regardless of whether this box included another step or
not. Now, the box is truncated if it reaches another step.

I've also reorganised the extensive debugging facilities within smf_fix_steps.