smurf: Use a macro for smf_free
smf_free only really calls astFree so rather than calling a function
to call astFree we redefine smf_free to be a macro. This allows
AST to report a bad free at the correct line number in SMURF.

In the longer term the suggestion is to replace all smurf
malloc/realloc calls with AST equivalents although we would
still like to retain the ability to run valgrind and the
debugger without calling the AST memory routines.