ndg: Fix duplicate ancestor problem
Previously, ndg issued an error if a provenance structure contained two
entries for the same NDF, created by the same command at the same time,
in which each of the two entry specified a different list of parents.

But look at the following example:

$KAPPA_DIR/cmult in=fred1 out=temp1
$KAPPA_DIR/cadd in=temp1 out=temp2
$KAPPA_DIR/cmult in=fred2 out=temp1
$KAPPA_DIR/cadd in=temp1 out=temp3
$KAPPA_DIR/add in1=temp2 in2=temp3 out=temp4

If the NDFs are small, the output NDFs created by the two CMULT commands
could have the same time stamp in the provenance (which only measures
time to 1 second). So the two "temp1.sdf" files have the same name,
creator and date, but will - legitamately - have different parents.
Previously, this would have caused NDG to report an error. Consequently
I've done away with this test. When purging duplicated ancestors from the
provenance info, I now consider two NDFs to be the same if they have the
same name, command and date, and if all pairs of corresponding parents in
the two NDFs are also the same.