smurf: calculate tswcs in concatenated data properly when padding
Since a LUT is used for the time axis in the tswcs frameset, the
previous method of simply shifting the time origin was incorrect.
The only reason concatenated data time axes looked sort-of right
when padding was included was because it does a linear
extrapolation. Now we just place linearly-extrapolated rts_end
values into the JCMTState for padded regions, and re-create the
tswcs using smf_create_tswcs, which is itself a wrapper for
sc2store_timeWcs.