smurf: Ignore previously flagged jumps when fitting data to a new jumps
Previously, the least squares fits that are used to determine the data
levels before and after a correlated DC step were including data that had
previously been flagged as a jump (i.e. the samples where the DC level is
moving from the lower level to the higher level). This could cause some
steps that have previously been fixed to be re-introduced (with a
different step height).