smurf: some test code for alternative sample variance calc in maps
Per was looking into how the map sample variance is calculated. In
order to do his version we need to keep track of an additional sum,
the total of the mapweights^2 which requires an additional map-sized
buffer in smf_iteratemap and an extra parameter to smf_rebinmap1. To
turn this alternative method on, uncomment the line at the top of
smf_rebinmap1 that defines __SMF_REBINMAP__SAMPLE_STANDARD_DEVIATION.