smurf: smf_templateFit1 can optionally just fit gain (not offset)
I've added this feature since we potentially have many model
components with DC level as a free parameter (mean level in
the map, COM, DKS...). This level is completely unconstrained
and is also totally degenerate between models. So, you can turn
off offset fitting in the linear fitter... for now I've just done
this for DKS but we may want to do this in other places.