smurf: extra status check and fix quality indexing smf_fix_steps
The status check is needed in case smf_get_goodrange failed (like
for a short smfData that is pure padding or apodization).
The tstride for the quality index was also incorrect, but we didn't
notice because we normally run this routine on bolo-ordered data.
For time-ordered data with padding it would almost certainly fail
(i.e. in an sc2clean call).