smurf: add ast.zero_notlast to skip boundary condition on last iter
This is probably a useful feature for deep maps of point sources.
For this to work it currently requires a fixed number of iterations
since it decides not to apply the boundary condition if smf_iteratemap
reaches maxiter. When this feature is used, map quality bits are still
set to SMF__MAPQ_ZERO, but they are not used to mask the final image.