smurf: Calculate steptime from JCMTSTATE for SCUBA-2
SCUBA-2 files are single sequences so we can calculate the actual
steptime directly from the RTS_END information. The RTS is not
able to give a step time that is exactly the value requested
and for a 5ms request you seem to get around 5.6ms.

Not tested in weird files where the RTS_END information is bad
at the end of file.