smurf: Refactor extinction correction tau relationships
We now store the tau relation coefficients (the values that let you
convert a CSO tau to a filter tau) in an external file
($SMURF_DIR/smurf_extinction.def) containing the coefficients for
each filter. These are read and the keymap is passed down to the
low level routines that do the scaling.

This patch also changes the default scaling from the SCUBA values
to the values determined from the SCUBA-2 filters by Jessica
Dempsey in May 2010.