smurf: bump version to 1.1.0, sc2clean uses config files and cleans DKS
Now that dark squids are stored as smfData's it was easy to get sc2clean
to process them. Switching to a config file simplified this -- we simply
have an identical set of cleandk.* parameters. Had to fudge a quality
array and the header information to get smf_clean_data to run but
otherwise it all seems to work well. Since we have a significant interface
change for sc2clean increased the version number. Don't yet handle DKS
cleaning in the iterative map-maker. Also, config files with 450.* or
850.* will currently make sc2clean barf.