smurf: Add missing configuration file parameters to MAKEMAP and SC2CLEAN
There is a lot of duplication of documentation here since the parameters
in SC2CLEAN are all used in MAKEMAP. Additionally many of the SC2CLEAN
parameters are also available in the NOI and FLT models.

Also add references to the location of the defaults files and other
minor documentation fixes.