smurf: Smooth COM using a filter that avoids discontinuities at start and end
I have found cases where smoothing the common mode using smf_boxcar1D can
introduce discontinuities at the start and end. For instance, the very
first smoothed COM value is the mean of COM values that come later (since
there are by definition no COM values earlier than the first one). So if
COM begins with a positive slope, the smoothed COM at the start is
biassed up above the slope, to some non-zero positive value. This means
there is a sudden leap at the start from zero to the non-zero value. When
COM is subtracted from the residuals, this leap gets transferred to the
residuals, which then causes the FLT model to introduce bad ringing.

The smf_boxcar1D function attempts to get round this by reducing the size
of the filter box at the start and end of the array. But this doesn't
work because the actual good COM signal does not start at the start of
the array - there is block of bad values first (the padded/apodised
region). By the time the good COM signal starts, the smf_boxcar1D filter
box has assumed it's full size, and so introduces the step mentioned above.

I've changed smf_calcmodel_com to smooth COM using the new routine
smf_tophat1D instead of smf_boxcar1D. The new routine performs the same
job as smf_boxcar1D but takes care to avoid the bias produced by an
assymetric dustribution of bad values in the filter box (it's
consequently slower than smf_boxcar1D).