smurf: more intelligent defaults for sc2clean
If we turn on DKS cleaning (with dkclean=1) we will certainly want
to remove the average from them first, so set cleandk.order=0.

Also, since both the DKS and the bolo data are cleaned using the
same routine, we have two confusing/useless parameters cleandk.dkclean
(we clearly don't want to clean the dark squids with themselves!) and
cleandk.flagstat (we don't bother carrying around a quality array...
if we don't want to make a map out of data when the telescope was
stopped that is handled by the quality for the bolometers). So,
these are now separated and explicitly turned off for now...