smurf: Tweak SMO quality handling
Do not use SMF__Q_MOD because we are not really fitting the data.
Use SMF__Q_FIT throughout to be consistent. Also, in the unlikely
event that we have found a new bad value, set the quality rather
than setting the data to bad.

Use COM quality for now since we have no more bits but there
should not be a SMO bad quality at the moment.