smurf: Handle quality differently to variance and data
 * Replace "unsigned char" usage for quality type with smf_qual_t
 * Use SMF__QUALTYPE instead of SMF__UBYTE
 * Use VAL__BADQ instead of VAL__BADUB

Also no longer store the pointer to quality in the pntr[] part of
smfData. It's now in standalone as "smf_qual_t qual". I think
that all usage treating pntr[2] as a special case is now using
smfData.qual as a special case. A handful of routines needed
some reorganizing to make this work.

This patch also protects any attempt to map an NDF quality
array with any type other than _UBYTE so we can not yet
change smf_qual_t to be an "unsigned short". We will need
some extra code in place to copy quality rather than mallocing
it and we will need something to compress quality and
copy it back out.