ndg: Add extra debug info in for "duplicate parents" error message.
Sarah Graves says that ORAC-DR is issuing messages about input NDFs being
included twice within the provenance of an output NDF, but with different
sets of parents. It's tricky to debug within ORAC-DR, so I have made NDG
create an error log file when this error is reported, containing verbose
information about the state of NDG at the time. Hopefully this will allow
me to work out why this error is being reported.

The error log file is ~/ndg.log

I need to remember to remove (or disable) this when "the case is solv-ed".