smurf: Switch internal quality type to 16 bits
We need more than 8 bits in the map-maker quality so we now copy
QUALITY from the NDF (if present) and store it in 16bits. If we
are writing out the quality we copy the quality back down to
8 bits. There are two methods:

 - if we are only using 8 bits we copy the bits we are using
 - if we are using more than 8 bits we compress related quality
   into a single entry.

The latter option is not reversible but what choice do we have
without switching NDF to 16 bit quality.