smurf: Allow smf_fix_steps to return a description of each fixed step
This is preparatory to creating a regression test facility for step
fixing. The step fixing algorithm is pretty fragile - making it work for
one data set often causes it to break for a different data set. So we
need some way of knowing that new changes do not break things for other
data sets.