smurf: Use sidequal everywhere in the map-maker
Previously most routines used by the map-maker had two arguments
for the data. The first was the smfData to be analyzed or modified
and the second a pointer to a quality array. This was done so that
the map-maker could share a single quality amongst multiple models.

This led to some problems though:

 1. The quality didn't know its own data order so there was no
    guarantee that the smfData and quality would match
 2. There was API bloat
 3. Each function had its own logic for deciding which pointer
    to use for quality.

This patch changes every API to use a single smfData argument.
The quality is stored in the sidequal slot of the smfData and
smf_select_qualpntr is used to determine which pointer should
be used. smf_dataOrder will already reorder everything to ensure
that the order matches.

This is a major code reorganization so I'm bound to have made
a mistaked somewhere. Maps do actually get made which is a good
start. Some of the models may now be extracting a qua_data when
they no longer need to so that will need looking at.

Of note is that smf_model_create no longer calculates the stationary
data. I couldn't work out why it was done there because
smf_clean_smfData already does it. If this is a mistake it's easy
to put it back in. Also smf_write_smfData will now always write
out the quality.