DATACUBE: Add pvslice script.
This script is being moved from its brief home in KAPPA.  DATACUBE is
the natural home in terms of functionality and style.  The script
doesn't permit full control of the plotting (like scaling and image
plane) but can readily be edited by a user in keeping with the other
DATACUBE scripts that are intended as examples as well as working
scripts.

Increment to Version 1.3, and update the documentation.