smurf: Change heater calibration
Following a discussion with Dan Bintley we have decided to correct
the conversion factor used to convert heater value to current. We
are also correcting the heater resistances for s4a and s8d to new
values that are consistent with the corrected conversion factor.

These changes will change the flatfield calibration and therefore
lead to a new definition of Flux Conversion Factor. The downside
of this is that the embedded flatfields that were stored in each
file when it was observed will now be inconsistent with the newly
calculated FCF. Any maps made with flatfield ramps will be okay
because their flatfields are recalculated. Older data will need
to have new flatfields associated with them.