smurf: Use iterative approach to step fixing
The main change is that an iteration loop is now used when fixing steps.
The first iteration takes out the big steps, subsequent iterations take
out smaller steps.

Also, the gradient estimates that are used to find steps are now
high-pass filtered (using a median filter) in order to allow short
increases of gradient on a larger high gradient background to be detected
more effectively.

Both the above changes mean that step fixing is now significantly slower,
but the good news is that it has not yet been multi-threaded so there is
scope for a serious speed-up.

There have also been a good number of more minor tweaks to the algorithm.

Note, the default dcthresh value has been changed to 6.0 in
dimmconfig.lis and in the default files.

This code is not as I would like it to be! It still needs lots of tidying
up, rationalising and better commenting. But it seems to work, and I
should be doing findback things, so I'll leave tidying it up for another
day.