smurf: Do not include padding in quality statistics
The quality reporting routines have been modified to optionally
ignore padding. The motiviation for this is that the padding
statistics can sometimes confuse things a lot. For example
I was getting FILT quality setting 10% of all samples to
bad but most of them were in the padding. Ignoring the padding
resulted in a much saner report of 1.5% that is inline with
the summary of the report indicating that more data were
retained when using the new filtering without apodization.

MAKEMAP and SC2CLEAN currently disable padding. We can easily
add a debugging report that turns padding back on.