smurf: More robust handling of bad flatfield ramps
In some SRO data we have aborted flatfield ramps that include bad
values for the heater. This throws off the heater fitting routine
and since the data are bad anyway there is nothing to salvage.

Rather than segv we now set status to SMF__BADFLAT and
smf_find_science skips the flatfield. Add additional status
and pointer definedness traps to allow calcflat and makemap
to complete when this happens.