KAPLIBS: Add new statistics routine to supersede KPG1_STATx.
The existing KPG1_STATx routine uses the standard one-pass
formula for the variance.

    Sum (x_i)**2 - (Sum x_i)**2/n
    -----------------------------
               n - 1

It has the disadvantage that it can be unstable to rounding errors for
large datasets differencing two large and similar sums, even though we
do compute in double precision, and there is no correction applied
such as that of Bjorck.

The simple two-pass algorithm while more robust, it can fall foul of
concentrated data with extreme outliers, and rounding, and is less
efficient too.

For an entire population the denominator be the population size n; it
is n-1 for a sample of observations.  What is appropriate depends on
the use being made.  For decent n, such as normal use in KAPPA:STATS,
it should make little practical difference.

Since we needed skewness and kurtosis, I've written a new routine that
uses Welford's recursive approach, extended to higher moments by
Terriberry.  It's a one-pass algorithm and more stable than before.

For consistency and ease of interpretation I've stuck with population
statistics.  If users want sample statistics they can apply
appropriate bias corrections.  For kurtosis the corrections seem to
vary in how to scale the 3 offset to derive excess kurtosis over pure
kurtosis, and whether or not there is another term.  Users can pick
their favourite.