smurf: Allow trailing flat ramp to be used if no previous ramp
Previously only the previous ramp in an observation was used.
Sometimes the previous ramp fails for whatever reason but the
end ramp works so it seemed silly to ignore that. Now we choose
the following ramp if there was no previous and if it was a ramp
that immediately followed the science data.