smurf: Add subarray enum
The FTS-2 code needs to know which subarray is being processed because
the FTS-2 is only in front of 4 of the subarrays. In order to make the
code easier to read Josh Oba had defined a private FTS-2 enum definition.

This patch moves that definition to sc2ast and now all of SMURF can
use it consistently. The new type is sc2ast_subarray_t.