dipso: cancel out equal, inverted, non-monotonic lutmaps when reading WCS from an NDF
For instance, the sp0wr command creates such Mappings, where the
(non-monotonic) AXIS array LutMap is in series with its own invert. Each
of these mappings is only defined in one direction, and so putting them
back to back causes the total mapping to be defined in neither direction.
Simplifying the total mapping causes the LutMaps to cancel out and be
removed, resulting in the Mapping being defined in both directions.
Reported by Ian Howarth.