KAPPA: Add Gapped histogram MODE to (C)LINPLOT.
As I suggested to Remo I thought an additional mode was better than
altering the existing "Histogram" behaviour.  MODE = "Gapped" works
like the "Histogram" mode, except vertical lines are not drawn either
side of bad values, thus there are just gaps where bad values are
located.  This was duplicated in CLINPLOT as its plots are small and
the vertical lines to the base are even more likely to be confused
with spikes.