KAPLIBS: Retrieve unintentional commits of non-split hunks.
This is the version of KPG1_PLOT I thought I had committed last night.
To restore where we should have been, withdraw the temporary
workaround to the KAPPA interface file.

A couple of hunks were not split on the second or third iteration of
"git add -p" to remove trailing blanks, and these included the part of
the TEMPSTYLE changes.  Apologies for the carelessness.  The plan was
to hold back on the addition of TEMPSTYLE in KPG1_PLOT, because after
delving deeper I discovered it was not part of the HISTOGRAM fix, and
if applied it would break several heavily used KAPPA applications.
Also we had not discussed the wider use of TEMPSTYLE, whether or not
it was the right approach.  If so, what should we do in the likes of
CLINPLOT where there are two main style files, or other tasks where
there's a KEYSTYLE.  Another route might be to include a special
character or keyword in a style group to indicate that style attributes
following it only apply to the current invocation.  That way to you
wouldn't need mirror TEMP<whatever>STYLE parameters for every
<whatever>STYLE parameter.

When I started the HISTOGRAM TEMPSTYLE addition, it was not
immediately clear which subroutines needed changing, as STYLE doesn't
appear in the code of histogram.f, unlike in linplot.f (appearing in
KPG1_ASSET and KPG1_PLTLN calls).  I shall continue to comment in the
application code where certain fixed-name parameters are defined to
help with maintenance.