KAPPA: Adding TEMPSTYLE to NORMALIZE, PSF, and SCATTER.
The introduction of TEMPSTYLE parameter to HISTOGRAM has the
ramification of adding it to NORMALIZE, PSF, and SCATTER because they
all use KPG1_GRPHW via KPG1_GRAPH.  I shall omit documenting this for
now in case we can find a programmatic route in the Parameter System
that avoids the use of an additional parameter for each style parameter.

At present it's only LINPLOT and recently HISTOGRAM that has
TEMPSTYLE, so it's not too late to put the genie back in the bottle
(although we may have to retain TEMPSTYLE in LINPLOT for backwards
compatibility).