smurf: now have fairly accurate memory estimate for common usage
More updates to smf_checkmem_dimm.

-temp space for initial data read (uses new maxfilelen parameter)
-added fudge factor for static model memory usage (empirical)
-temp space rough estimate for data cleaning (empirical)

A fair bit of trial and error went into this. Perhaps the biggest
surprise was that data cleaning requires about 3*full data cube of
temporary space, so we should look into that. I also had to increase the
size estimate for static model containers by about 20% over everything I
could think of. The dynamic model containers (i.e. not RES,AST,QUA,LUT)
seem to be fairly accurate.

This estimate is quite close to reality for the tests I did. We already
manufacture a margin of error into the "maximum available memory" which
is 80% of the physical memory in smurf_makemap, so I didn't think we
should add more in here unnecessarily.