CONVERT: Fix two bugs coping with output from FTGTDM for _CHAR vectors.
Set the dimension of a vector character array to the previously
calculated number of elements given by the repeat count / width.
FTGTDM is used to define the dimensions stored in the TDIMn keyword.
The first bug manifests when TDIMn is absent, it's a _CHAR vector, and
annoyingly FTDTDM returns the repeat count.  I'm not sure if this has
always been true or was changed for C programmers.

Correct AJC's 2002 March 13 fix to handle vector _CHAR arrays.  While
there is no need to use TDIMn for a one-dimensional character
array---most writers would use just TFORMn = 'rAw'---it is part of the
standard.  The FTGTDM documentation does not mention that the returned
dimensions for a _CHAR array are (width, dim1 [, dim2, ...]).  It is
described elsewhere, which presumably explains why I failed to include
this special case in the original version.