smurf: clean data before opening models to lower memory requirements
There is a spike in memory usage during data cleaning. We can
lower its impact by cleaning data when other models are not
currently in memory.

For memiter=1 (default) the cleaning now happens immediately
after concatenation.

For memiter=0 it happens as before within the iteration loop, except
smf_clean_smfData is now called as soon as RES and QUA are loaded
(and the other models are loaded afterwards).

There also appeared to be a bug in the memiter=0 case in which the
sidecar quality was only being associated with RES and AST for
the first iteration. Since the files are continually re-loaded
after each iteration, the association has been moved out of the
"if(iter==0)" clause.