KAPLIBS: Remove or replace the remainder of the prefix-less routines.
Create new subroutine KPG_FISEx like KPG1_FILLx but applies
only to a region define by pixel bounds.  BAD2Dx is gone.

ICMMM replaced by generic and improved KPG_IMMMx.  The description
about how the iterations cease is not quite correct, in that there's
no test for convergence, but it is what the code should do,
and I'll modify the code to accordingly in due course.