smurf: smf_deepcopy_smfData can re-order
The purpose was to avoid a potential memory usage spike in
smf_fft_data since it calls smf_dataOrder immediately after
a copy. I now realize this case doesn't happen
in practice. Regardless, this may be useful in other places.

I factored the private smf__dataOrder_array out of smf_dataOrder
so that it can be called in smf_deepcopy_smfData.