KAPLIBS: Improve median derivation to yield more-robust mode in KPG_IMMMx.
Uses the central 20% of the data for the median averaging after
rejecting outliers or the previous nine or ten central values if
that's larger.  The goal is to obtain a more-stable median for
deriving the mode.  Add anther semi-arbitrary tolerance test to exit
iterations, although I doubt this will be commonly reached.

Corrected the termination description.