smurf: add ast.zero_circle map constraints
This is similar to ast.zero_lowhits, except now you can
specify the RA, Dec and radius of a circle beyond which you
would like to constrain the map to 0 after each iteration.
This constraint will obey ast.zero_notlast if you wish to
suppress the constraint on the last iteration.