smurf: add freeOld flag to smf_dataOrder_array
This fixes a bug introduced in 1fe40813883f2ac2bc61445da9aea8222c694d34
when running SC2CLEAN. SC2CLEAN seems to be the only place that
re-ordering copy is ever used. The reason it broke is that
smf_dataOrder_array frees the "old" data buffer before returning (unless
doing an in-place re-order). In the case of copying, you don't want to
free the reference buffer.