smurf: Add specialist pointing configs
These are like the bright_compact and bright_extended
configs except that they use the non-apodizing filter
and the compact one does not use com.notfirst.

It's possible that when we become happier with the non-apodising
filter these configs will disappear and we will just use the
standard bright variants.