smurf: Change the test for procimity to a source when fixing steps.
The edges of a bright source can trigger the step corrector. So steps
close to a bright source are ignored. Sources used to be detected by
looking for flat section in the median smoothed bolometer data, but this
is unreliable as other things can cause flat section. So now tsource are
detected by comparing the largest and smallest data values in a region
close to the step.