smurf: despite heavy commenting, still got confused about chunks...
The arrays of smfArrays are meant to coincide with each file
loaded in for the memiter=0 case. When iterating in memory, there
is only 1 smfArray for each model! (hence subscript [0] hard-wired).
The only reason contchunk is passed to smf_write_bolomap/shortmap
is to get the name of the extensions right.