smurf: implement FILT_EDGE_SMALLSCALE/LARGESCALE + scanvel bug fixes
To specify high and low-pass edge filters you can now specify
FILT_EDGE_SMALLSCALE and FILT_EDGE_LARGESCALE (both for pre-processing
and in the FLT model). It will translate these scales to override
FILT_EDGEHIGH and FILT_EDGELOW based on a measurement of the
mean scan velocity.

Also fixed up a number of bugs in my initial attempt to load in the
SCAN_VEL FITS header, and in the measured value from smf_flag_slewspeed.