smurf: Change MCEHEAD writing for sc2store
Previously we stored an integer array in an MCEHEADS extension
containing a bunch of integers for every time slice. This array
was mainly the same numbers repeated ntslice times or else
zeroes.

We have changed things so that a single MCE header is written
out in an MCEHEAD extension. We populate this from an AstKeyMap
and use the kaplibs kpg1Ky2hd routine to write the keymap
to an HDS extension.