smurf: Fix bug in step correction for very short scans
For very short scans the DCMAXSTEPS requirement can cause a bolometer to
be rejected even if it has just a single step. To fix this, I've imposed
a hard lower limit of four steps in a bolometer before the bolometer is
rejected, no matter how short the scan is. For longer scans, the DCMAXSTEPS
parameter takes over.