smurf: Addd option to pad start and end with artifical data rather than zeros
A new config parameter ZEROPAD has been added. If ZEROPAD is 1 (the
default), the start and end of each time stream is padded with zeros.
Otherwise, they are padded with artificial data. This is a cubic curve to
which Gaussian noise is added. The cubic is chosen so that the data is
continuous in an FFT wrap-around sort of way.

The ideas is that with ZEROPAD set to 0, it should not be necessary to
apodise the data.