smurf: Change cubic padding to limited padded values
This change gets round the problem where an unusallay high gradient at
the start or end of the time series could cause the cubic interpolation
to reach silly high values.

The gradient at the end of each bolometer time series is estimated by
fitting a straight line to the last 100 samples. The following 100
samples of padding is a quadratic curve that rolls the gradient off
linearly to zero (with added noise). Likewise, the 100 samples of padding
that preceeds the start of the time series is a quadratic curve that
starts with gradient zero and ends with the gradient estimated from the
first 100 samples of the time stream. A cubic curve is then used to
connect the two quadratic curves (in a wrap-around sense).