smurf: update handling of convergence to enable proper usage of ast.zero_notlast
The way things were previously set up, if we were using a variable number of
iterations (i.e. iterate to convergence), we would not know in advance when
calling smf_calcmodel_ast whether it was the last pass through the loop
or not. This means that we couldn't properly use out "ast.zero_notlast" flag.
Now it checks for convergence, and then does one final pass (unless of
course that would exceed the maximum).