smurf: Handle common quality states with special code
When counting quality bits we know that if the quality is 0 then
no bits will be set. We also know that BADDA+BADBOL is a very
common state so we check that quickly before doing the loop
over each bit.

Can make smf_qualstats 4 times faster depending on how many
quality bits are set.