KAPLIBS: Add KPG1_GTGPT, a variant of KPG1_GTGRP that handles temporary attributes.
For the moment this seems to be working on several examples, including
where it is invoked twice for the same parameter (e.g. in MLINPLOT).
It has been this additional requirement that took most time to
resolve. It's a bit baroque passing the control information to
indicate that KPG1_GTGPT will be called again via a suffix to the
parameter name, but it's the only way I could devise that actually
worked.

I admit that the uncertainty whether SUBPAR_CURSAV or SUBPAR_PUTNAME
is needed in different circumstances leaves me concerned that holes
will appear in the dyke wall as users try to break it (unintentionally
of course), and we saw what happened in Hungary.  [Before Alan
retired, I thought he was going to provide programmer manuals for
important system libraries like SUBPAR, and the fixed part...] I
suspect that when David switched these two SUBPAR calls in KPG1_GTGRP,
the new code worked, but in other circumstances the old code writes
the value successfully.