KAPLIBS: Incorporate temporary-attribute facilities into KPG1_ASSET.
This is via a baroque scheme but which avoids changing the KPG1_ASSET
API.  The parameter name may have a non-alphanumeric prefix to
indicate that temporary attributes should be parsed.  There is also an
optional suffix to request that the full expression be retained in the
parameter file until the final invocation of KPG1_GTGPT that records
just the persistent attributes.  Otherwise some plotting style
information may not be acted upon in the second and later invocations.
It might be better to minimise the tricky syntax and strip this suffix
from being passed to KPG1_GTGPT and use an additional logical
argument.