smurf: initial commit of whitening filter
smf_whiten fits a power-law to the low-frequency end of the data,
adds it to a flat value that is fit to a higher-frequency portion
of the spectrum to estimate the white noise level, and then divides
by this function to create a nearly flat power spectrum.

It is currently called within smf_filter_exectute. There is a
new configuration parameter called "whiten" which, when set, will
cause whitening to be applied as a pre-processing step (it is
therefore also available to sc2clean). smf_filter_execute was
modified slightly so that if an empty filter is supplied, it
will still run and only apply the whitening factors.

To see what it does, use sc2concat to combine some data, run
it through sc2clean with default parameters, and a second
time setting "whiten=1". The latter will have essentially no
large baseline drifts. You can also run both files through
sc2fft with "power=true" set, and you will see how the
low-frequency end has been flattened out. The only remaining
features are lines in the power spectrum (some narrow, and
some quite wide), as well as the anti-aliasing filter which
begins to noticeably roll-off the power spectra above about
50 Hz.